rbd: replace deprecated rbd_list() with rbd_list2()
